+++ tests/Favorite/FavoriteServiceTest.php
@@ -26,49 +26,49 @@ final class FavoriteServiceTest extends ApiTestCase
         parent::setUp();
         $this->favService = $this->buildFavoriteService();
 
-        $this->favService->addDeclinationToUserFavorites(1);
+        $this->favService->addDeclinationToUserFavorites('1');
     }
 
     public function testAddProductToFavorite()
     {
-        $this->favService->addDeclinationToUserFavorites(2);
+        $this->favService->addDeclinationToUserFavorites('2');
 
-        $this->assertTrue($this->favService->isInFavorites(2));
+        $this->assertTrue($this->favService->isInFavorites('2'));
     }
 
     public function testAddFavProductToFavorite()
     {
         $this->expectException(FavoriteAlreadyExist::class);
 
-        $this->favService->addDeclinationToUserFavorites(1);
+        $this->favService->addDeclinationToUserFavorites('1');
     }
 
     public function testAddNotProductToFavorite()
     {
         $this->expectException(CannotFavoriteDisabledOrInexistentDeclination::class);
 
-        $this->favService->addDeclinationToUserFavorites(404);
+        $this->favService->addDeclinationToUserFavorites('404');
     }
 
     public function testIsNotFavorite()
     {
-        $isFavorite = $this->favService->isInFavorites(3);
+        $isFavorite = $this->favService->isInFavorites('3');
 
         $this->assertFalse($isFavorite);
     }
 
     public function testIsFavorite()
     {
-        $isFavorite = $this->favService->isInFavorites(1);
+        $isFavorite = $this->favService->isInFavorites('1');
 
         $this->assertTrue($isFavorite);
     }
 
     public function testGetAll()
     {
-        $this->favService->addDeclinationToUserFavorites(2);
+        $this->favService->addDeclinationToUserFavorites('2');
         $favorites = $this->favService->getAll();
-        $this->favService->removeDeclinationToUserFavorites(2);
+        $this->favService->removeDeclinationToUserFavorites('2');
 
         $this->assertTrue(is_array($favorites));
         $this->assertCount(2, $favorites);
@@ -80,15 +80,15 @@ final class FavoriteServiceTest extends ApiTestCase
 
     public function testRemoveProductFromFavorite()
     {
-        $this->favService->removeDeclinationToUserFavorites(1);
+        $this->favService->removeDeclinationToUserFavorites('1');
 
-        $this->assertFalse($this->favService->isInFavorites(1));
+        $this->assertFalse($this->favService->isInFavorites('1'));
     }
 
     public function tearDown() :void
     {
-        $this->favService->removeDeclinationToUserFavorites(1);
-        $this->favService->removeDeclinationToUserFavorites(2);
+        $this->favService->removeDeclinationToUserFavorites('1');
+        $this->favService->removeDeclinationToUserFavorites('2');
         parent::tearDown();
     }